Fixed favorites and make sorting with id
The by_index function is remained for convinience, removing it should not cause a problem.
Copied from my steam post about the fav problem:
The reason I think that the original code did not work is because of the line
[code] all_tools[index].favorite = true [/code]
and specifically the all_tools[index] part. The index variable here is referring to the index value of the tool but for the all_tools table, the index is the order of the items that will be displayed. As during the sorting of generate_all_tools(), the index are changed.
My modified code does as below:
Iterate through every tool in game and see if they are on the list of favorite (read from persistant), if they are flag them.
So the problem with index is mitigated.
However this may be not efficient at all and since I am not very deep into lua programming and theories I do not have an idea of how to improve this.